pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pxinv"
version = "0.1.0"
description = "Proxmox VM & container inventory CLI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["proxmox", "homelab", "cli", "inventory", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "proxmoxer>=2.0",
    "requests>=2.28",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
pxinv = "pxinv.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/pxinv"
Issues = "https://github.com/yourusername/pxinv/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "pip-audit>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pxinv*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"